data_common/scripts/requirements.txt
pandas
geopandas
python-dotenv
SQLAlchemy
psycopg2-binary
